Conflicts:
	log/src/main/java/de/cubeisland/engine/log/action/logaction/kill/KillActionType.java
	log/src/main/java/de/cubeisland/engine/log/storage/LogEntry.java

+++ basics/src/main/java/de/cubeisland/engine/basics/storage/Mail.java
@@ -54,7 +54,7 @@ public class Mail extends UpdatableRecordImpl<Mail> implements Record4<UInteger,
         {
             return ChatFormat.RED + "CONSOLE" + ChatFormat.WHITE + ": " + this.getMessage();
         }
-        User user = CubeEngine.getUserManager().getUser(this.getSenderid().longValue());
+        User user = CubeEngine.getUserManager().getUser(this.getSenderid());
         return ChatFormat.DARK_GREEN + user.getName() + ChatFormat.WHITE + ": " + this.getMessage();
     }
